requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
langchain-ollama
pypdf
langchain-groq
python-dotenv
streamlit
